add last_id to users table for non-message user getters
If a message was signed with a differently-blinded session_id than the
one in the users table for a user, that session_id goes in the alt_id
column of the message.  However, when fetching a user list (e.g.
moderators for a room), it is useful for now to give the session_id
that user last used.